Update Administration to V2.0.0.

+++ core/changeprefix.php
@@ -1,15 +1,31 @@
 <?php
-/*
- * LibreCMS - Copyright (C) Diemen Design 2018
- * This software may be modified and distributed under the terms
- * of the MIT license (http://opensource.org/licenses/MIT).
+/**
+ * LibreCMS - Copyright (C) Diemen Design 2019
+ *
+ * Core - Change Database Tables Prefix
+ *
+ * changeprefix.php version 2.0.0
+ *
+ * LICENSE: This source file may be modifired and distributed under the terms of
+ * the MIT license that is available through the world-wide-web at the following
+ * URI: http://opensource.org/licenses/MIT.  If you did not receive a copy of
+ * the MIT License and are unable to obtain it through the web, please
+ * check the root folder of the project for a copy.
+ *
+ * @category   Administration - Core - Change Database Tables Prefix
+ * @package    core/changeprefix.php
+ * @author     Dennis Suitters <dennis@diemen.design>
+ * @copyright  2014-2019 Diemen Design
+ * @license    http://opensource.org/licenses/MIT  MIT License
+ * @version    2.0.0
+ * @link       https://github.com/DiemenDesign/LibreCMS
+ * @notes      This PHP Script is designed to be executed using PHP 7+
  */
-echo'<script>/*<![CDATA[*/';
-if(session_status()==PHP_SESSION_NONE)
-  session_start();
+echo'<script>';
+if(session_status()==PHP_SESSION_NONE)session_start();
 $dbprefix=isset($_POST['dbprefix'])?filter_input(INPUT_POST,'dbprefix',FILTER_SANITIZE_STRING):'';
 $dbprefix=trim($dbprefix);
-require_once'db.php';
+require'db.php';
 if($settings['database']['prefix']!=$dbprefix){
   $result=$db->query("SHOW TABLES FROM `".$settings['database']['schema']."` LIKE '%".$settings['database']['prefix']."%'");
   $renamed=$failed=0;
@@ -42,12 +58,11 @@ if($settings['database']['prefix']!=$dbprefix){
        'version = '.time().PHP_EOL.
        'url = '.$settings['system']['url'].PHP_EOL.
        'admin = '.$settings['system']['admin'].PHP_EOL;
-  if(file_exists('config.ini'))
-    unlink('config.ini');
+  if(file_exists('config.ini'))unlink('config.ini');
   $oFH=fopen("config.ini",'w');
   fwrite($oFH,$txt);
   fclose($oFH);
   echo'window.top.window.$.notify({type:"info",icon:"",message:"'.$error.'});';
 }else
   echo'window.top.window.$.notify({type:"danger",icon:"",message:"Tables are already Prefixed with `'.$dbprefix.'`"});';
-echo'window.top.window.$("#blocker").remove();/*]]>*/</script>';
+echo'window.top.window.$("#blocker").remove();</script>';
